refactor: remove unused components
