--- app/src/components/molecules/search/search.styled.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Search = styled.div`
-  width: 100%;
-`;
