+++ app/src/components/organisms/filters/filters.styled.ts
@@ -3,6 +3,5 @@ import styled from "styled-components";
 export const Filters = styled.form`
   display: flex;
   flex-direction: column;
-  justify-content: center;
   row-gap: 8px;
 `;
