--- app/src/components/templates/list/list.styled.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const List = styled.div`
-  width: 70%;
-`;
