--- app/src/components/templates/sidebar/sidebar.styled.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const Sidebar = styled.div`
-  width: 30%;
-  display: flex;
-  flex-direction: column;
-  padding: 24px;
-  box-sizing: border-box;
-`;
